# genproto/server/v1/leaderboards.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "server/v1/leaderboards.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "LeaderboardsService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/server/leaderboards/{board_id}/entries/{subject_id}": {
      "get": {
        "operationId": "LeaderboardsService_GetLeaderboardEntry",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1LeaderboardScoreSnapshot"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/v1ErrorResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "board_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "subject_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "LeaderboardsService"
        ]
      }
    },
    "/v1/server/leaderboards/{board_id}/top": {
      "get": {
        "operationId": "LeaderboardsService_ListLeaderboardTop",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListLeaderboardTopResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/v1ErrorResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "board_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "page_token",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "LeaderboardsService"
        ]
      }
    },
    "/v1/server/leaderboards/{board_id}:submit": {
      "post": {
        "operationId": "LeaderboardsService_SubmitLeaderboardScore",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1LeaderboardScoreSnapshot"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/v1ErrorResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "board_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/torchwoodserverv1LeaderboardsServiceSubmitLeaderboardScoreBody"
            }
          }
        ],
        "tags": [
          "LeaderboardsService"
        ]
      }
    }
  },
  "definitions": {
    "torchwoodserverv1LeaderboardsServiceSubmitLeaderboardScoreBody": {
      "type": "object",
      "properties": {
        "subject_id": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "int64"
        },
        "tiebreak_value": {
          "type": "string",
          "format": "int64"
        },
        "period": {
          "type": "string",
          "title": "缺省 = 当前期；显式必须命中 {当前期, 上一期}（离线补传宽限）"
        },
        "request_id": {
          "type": "string"
        }
      },
      "title": "SubmitLeaderboardScoreRequest：server 面可代任意 subject 提交（函数/\n运营/自动化）。request_id 走 (project, actor, request_id) 24h 幂等——\n网络层重试不重复烧 per-subject 限频额度。"
    },
    "v1Error": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "code": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "error_id": {
          "type": "string"
        },
        "error_code": {
          "$ref": "#/definitions/v1ErrorCode"
        }
      }
    },
    "v1ErrorCode": {
      "type": "string",
      "enum": [
        "ERROR_CODE_UNSPECIFIED",
        "ERROR_CODE_INVALID_REQUEST",
        "ERROR_CODE_RESOURCE_NOT_FOUND",
        "ERROR_CODE_INVALID_CREDENTIALS",
        "ERROR_CODE_PERMISSION_DENIED",
        "ERROR_CODE_RESOURCE_CONFLICT",
        "ERROR_CODE_QUOTA_EXCEEDED",
        "ERROR_CODE_PRECONDITION_FAILED",
        "ERROR_CODE_CONCURRENT_MODIFICATION",
        "ERROR_CODE_VALUE_OUT_OF_RANGE",
        "ERROR_CODE_OPERATION_NOT_ALLOWED",
        "ERROR_CODE_INTERNAL_ERROR",
        "ERROR_CODE_SERVICE_UNAVAILABLE",
        "ERROR_CODE_TIMEOUT"
      ],
      "default": "ERROR_CODE_UNSPECIFIED",
      "title": "ErrorCode 与 gRPC code 的映射（HTTPErrorHandler 单一事实来源，\n见 internal/infra/server/errors.go）：\n  - InvalidArgument → ERROR_CODE_INVALID_REQUEST\n  - FailedPrecondition → ERROR_CODE_PRECONDITION_FAILED\n  - NotFound → ERROR_CODE_RESOURCE_NOT_FOUND\n  - AlreadyExists → ERROR_CODE_RESOURCE_CONFLICT\n  - Aborted → ERROR_CODE_CONCURRENT_MODIFICATION\n  - Unauthenticated → ERROR_CODE_INVALID_CREDENTIALS\n  - PermissionDenied → ERROR_CODE_PERMISSION_DENIED\n  - ResourceExhausted → ERROR_CODE_QUOTA_EXCEEDED\n  - DeadlineExceeded → ERROR_CODE_TIMEOUT\n  - 其余 → ERROR_CODE_INTERNAL_ERROR"
    },
    "v1ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/v1Error"
        }
      }
    },
    "v1LeaderboardEntry": {
      "type": "object",
      "properties": {
        "board_id": {
          "type": "string"
        },
        "period": {
          "type": "string"
        },
        "subject_id": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "int64"
        },
        "tiebreak_value": {
          "type": "string",
          "format": "int64"
        },
        "submit_count": {
          "type": "integer",
          "format": "int32"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "LeaderboardEntry 是排行榜条目：唯一键 (board, period, subject) 内建去重。"
    },
    "v1LeaderboardScoreSnapshot": {
      "type": "object",
      "properties": {
        "board_id": {
          "type": "string"
        },
        "period": {
          "type": "string"
        },
        "total": {
          "type": "string",
          "format": "int64"
        },
        "entry": {
          "$ref": "#/definitions/v1LeaderboardEntry"
        },
        "rank": {
          "type": "string",
          "format": "int64"
        },
        "position": {
          "type": "string",
          "format": "int64"
        },
        "below": {
          "type": "string",
          "format": "int64"
        }
      },
      "title": "LeaderboardScoreSnapshot 是 submit / me / getEntry 的统一响应：\nrank = competition（并列同名次）；position = 全序位次（tie-break 决出）；\nbelow = value 严格低于我的人数（百分位分子，不受 tie-break 影响）。\n无条目时 entry/rank/position/below 缺省，total 仍返回。"
    },
    "v1LeaderboardTopEntry": {
      "type": "object",
      "properties": {
        "subject_id": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "int64"
        },
        "tiebreak_value": {
          "type": "string",
          "format": "int64"
        },
        "rank": {
          "type": "string",
          "format": "int64"
        },
        "position": {
          "type": "string",
          "format": "int64"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "LeaderboardTopEntry 是 top 列表行。"
    },
    "v1ListLeaderboardTopResponse": {
      "type": "object",
      "properties": {
        "period": {
          "type": "string"
        },
        "total": {
          "type": "string",
          "format": "int64"
        },
        "entries": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1LeaderboardTopEntry"
          }
        },
        "next_page_token": {
          "type": "string"
        }
      },
      "title": "ListLeaderboardTopResponse 是 top 分页响应。"
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "description": "格式: Bearer \u003cjwt\u003e",
      "name": "Authorization",
      "in": "header"
    },
    "apiKey": {
      "type": "apiKey",
      "description": "项目绑定在密钥行上，不需要 X-Torchwood-Project。该头仅 Console admin 会话切换项目。",
      "name": "X-API-Key",
      "in": "header"
    },
    "cookie": {
      "type": "apiKey",
      "description": "TORCHWOOD_session_console=\u003csid\u003e（Console admin 会话）",
      "name": "Cookie",
      "in": "header"
    }
  },
  "security": [
    {
      "apiKey": []
    }
  ],
  "x-torchwood-access": "server"
}